+++ packages/browser-extension/package.json
@@ -50,7 +50,6 @@
     "react-art": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-native-gesture-handler": "^2.24.0",
-    "react-native-pager-view": "^6.7.0",
     "react-native-paper": "^5.12.5",
     "react-native-paper-tabs": "^0.11.3",
     "react-native-reanimated": "^3.17.0",
