+++ packages/common/src/index.ts
@@ -1,5 +1,6 @@
 export * from "./conclusions";
 export * from "./entities";
+export * from "./entityHistory";
 export * from "./logging";
 export * from "./outcomes";
 export * from "./urls";
